refactor(monitoring): testability and resilience
